Styles avec DaisyUI

+++ client/src/components/PokemonListing.tsx
@@ -29,12 +29,14 @@ export default function PokemonListing({
   });
 
   const handleDelete = async (id: number) => {
+    if (!confirm("Êtes-vous sûr de vouloir supprimer ce Pokémon ?")) {
+      return;
+    }
+
     const res = await deletePokemon(id);
 
     const newData = data.filter((pokemon) => pokemon.id !== id);
     setData(newData);
-
-    alert(res.message);
   };
 
   if (filteredData.length === 0) {
@@ -46,7 +48,7 @@ export default function PokemonListing({
   }
 
   return (
-    <table>
+    <table className="table table-compact w-full">
       <thead>
         <tr>
           <th>#</th>
@@ -58,11 +60,12 @@ export default function PokemonListing({
           <th>Attaque Spé.</th>
           <th>Défense Spé.</th>
           <th>Vitesse</th>
+          <th></th>
         </tr>
       </thead>
       <tbody>
         {filteredData.map((pokemon) => (
-          <tr key={pokemon.id}>
+          <tr key={pokemon.id} className="hover">
             <td>{pokemon.id}</td>
             <td className={styles[pokemon.type[0].toLowerCase()]}>
               {pokemon.name[currentLang]
@@ -81,7 +84,25 @@ export default function PokemonListing({
             ))}
 
             <td>
-              <button onClick={() => handleDelete(pokemon.id)}>X</button>
+              <button
+                onClick={() => handleDelete(pokemon.id)}
+                className="btn btn-circle btn-error btn-sm"
+              >
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  className="h-6 w-6"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  stroke="currentColor"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth="2"
+                    d="M6 18L18 6M6 6l12 12"
+                  />
+                </svg>
+              </button>
             </td>
           </tr>
         ))}
